+++ src/AddTodoForm.jsx
@@ -1,0 +1,9 @@
+export default function AddTodoForm() {
+    return (
+        <form>
+            <label htmlFor="todoTitle">Title </label>
+            <input type="text" id="todoTitle"></input>
+            <button type="submit">Add</button>
+        </form>
+    )
+}
